Events service -> recordEvent function
