+++ client/app/services/events.js
@@ -1,30 +1,9 @@
-import { debounce, extend } from 'lodash';
-import { $http } from '@/services/ng';
+import recordEvent from '@/services/record-event';
 
 export let Events = null; // eslint-disable-line import/no-mutable-exports
 
-let events = [];
-
-const sendEvents = debounce(() => {
-  const eventsToSend = events;
-  events = [];
-
-  $http.post('api/events', eventsToSend);
-}, 1000);
-
 function EventsService() {
-  this.record = (action, objectType, objectId, additionalProperties) => {
-    const event = {
-      action,
-      object_type: objectType,
-      object_id: objectId,
-      timestamp: Date.now() / 1000.0,
-      screen_resolution: `${window.screen.width}x${window.screen.height}`,
-    };
-    extend(event, additionalProperties);
-    events.push(event);
-    sendEvents();
-  };
+  this.record = recordEvent;
 }
 
 export default function init(ngModule) {
